Add MoveCommand tests
Don't require subsystems to be online for MoveCommand or MoveAbsoluteCommen
command to run anymore as they now don't do anything in execute when
subsystem is offline.

Also remove usage of default parameters for general commands as that
doesn't make sense anymore as command could be used for a wide variety
of subsystems and thus have wide variety of practical defaults.